requirements.txt
instaloader>=4.10
flask>=3.0
flask-sqlalchemy>=3.1
sqlalchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13
gunicorn>=21.0
python-dotenv>=1.0
